backend/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
marshmallow
Flask-JWT-Extended
python-dotenv
pytest
